FIX: Only refresh travis list when it's actually different.
Refreshing travis every 2 minutes does a great job of keeping the dashboard up-to-date, but right now
it creates a lot of unnecessary flickering.

This patch blocks the send_event() command unless the items are actually different.  It uses a global
variable (gack) but it wasn't clear to me where a non-global variable would be placed.