API Removed big query support
The project lead (Ilya Grigorik) acknowledged the limitations as structural,
so we can't easily work around them. Quote:

"In retrospect, I wish I did partition the dataset into monthly or quarterly tables... but that's only with the benefit of hindsight. At the time when I was setting up the project, the decision to go with one big table was a conscious one - the worry was that it would make queries across larger date ranges more complicated. Knowing what I know today, I would have gone the other way. "